+++ app/javascript/components/Home.js
@@ -1,11 +1,6 @@
-import React, { useEffect } from 'react'
+import React from 'react'
 
 export default function Home() {
-
-
-  useEffect(() => {
-    console.log(greeting)
-  }, [greeting]);
 
   return (
     <div>
